whisper-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai-whisper==20231117
torch==2.5.1
